Removed owner pill and rended the tag inside the component
